+++ bee-block/src/protocol.rs
@@ -1,0 +1,29 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use packable::{prefix::StringPrefix, Packable};
+
+use crate::{output::RentStructure, Error};
+
+/// Defines the parameters of the protocol.
+#[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Packable)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[packable(unpack_error = Error)]
+pub struct ProtocolParameters {
+    // The version of the protocol running.
+    version: u8,
+    // The human friendly name of the network.
+    #[packable(unpack_error_with = |err| Error::InvalidNetworkName(err.into_item_err()))]
+    network_name: StringPrefix<u8>,
+    // The HRP prefix used for Bech32 addresses in the network.
+    #[packable(unpack_error_with = |err| Error::InvalidBech32Hrp(err.into_item_err()))]
+    bech32_hrp: StringPrefix<u8>,
+    // The minimum pow score of the network.
+    min_pow_score: u32,
+    // The below max depth parameter of the network.
+    below_max_depth: u8,
+    // The rent structure used by given node/network.
+    rent_structure: RentStructure,
+    // TokenSupply defines the current token supply on the network.
+    token_supply: u64,
+}